+++ packages/amplify-e2e-tests/src/aws-matchers/index.ts
@@ -1,0 +1,2 @@
+require('./s3matcher');
+require('./iamMatcher');