+++ packages/amplify-e2e-tests/src/init/index.ts
@@ -1,0 +1,8 @@
+export {
+  default as initProjectWithProfile,
+  initProjectWithAccessKey,
+  initNewEnvWithAccessKey,
+  initNewEnvWithProfile
+} from './initProjectHelper';
+export { getProjectMeta } from '../utils';
+export { default as deleteProject } from './deleteProject';