+++ packages/amplify-provider-awscloudformation/lib/push-resources.js
@@ -449,14 +449,16 @@ function formNestedStack(context, projectDetails, categoryName, resourceName, se
           Object.assign(parameters, { env: currentEnv });
         }
 
-        templateURL = resourceDetails.providerMetadata.s3TemplateURL;
-        nestedStack.Resources[resourceKey] = {
-          Type: 'AWS::CloudFormation::Stack',
-          Properties: {
-            TemplateURL: templateURL,
-            Parameters: parameters,
-          },
-        };
+        if (resourceDetails.providerMetadata) {
+          templateURL = resourceDetails.providerMetadata.s3TemplateURL;
+          nestedStack.Resources[resourceKey] = {
+            Type: 'AWS::CloudFormation::Stack',
+            Properties: {
+              TemplateURL: templateURL,
+              Parameters: parameters,
+            },
+          };
+        }
       }
     });
   });